+++ src/Commands/Provisioning/Tenant/InvokeTenantTemplate.cs
@@ -295,7 +295,7 @@ namespace PnP.PowerShell.Commands.Provisioning.Tenant
                 }
                 catch
                 {
-                    throw new PSInvalidOperationException($"Your template contains artifacts that require an access token for https://{Connection.GraphEndPoint}. Please provide consent to the PnP Management Shell application first by executing: Register-PnPManagementShellAccess");
+                    throw new PSInvalidOperationException($"Your template contains artifacts that require an access token for https://{Connection.GraphEndPoint}. Please provide consent to the EntraID application first by executing: Register-PnPEntraIDApp or Register-PnPEntraIDAppForInteractiveLogin");
                 }
             }